remove one _get_root_tables from hma
